Push down the way "bazel_tools" is declared
Instead of simply injecting "bazel_tools" into the module repo mapping,
make it so that we implicitly add a bazel_dep("bazel_tools") to all
MODULE.bazel files. This ensures that we also properly process the
MODULE.bazel file that is part of bazel_tools itself. This allows us to
get rid of some hacks on the bb-storage side.