Let CompatibleToolchainsForType call VisibleTarget on the platform
Right now we have some code to manually attempt to extract the target
platform the configuration. It turns out we don't need to do that if we
simply call VisibleTarget. VisibleTarget is already capable of expanding
label_flag()s. The advantage is that we get less repetition of code, and
also allow target platforms to be aliases.