Make bazel-lib platform_transition_filegroup() work
--platforms is interesting, in that Bazel models it as a repeated labels
setting, even though as part of a transition it can realistically only
be set to a single value. Solve this by adding a singleton_list option.

While there, add extra_execution_platforms. This needs to be a
label_list, which Bazel does not support for build settings. Add support
for this as an extension.