Re-add the ability to omit the default exec group
When I made ConfiguredTarget lazy, I removed support for this. This
seems to have been a mistake in retrospect, because it makes
configuration of targets sequential. Now that we want to make this eager
again, we should have this. Otherwise even the simplest rules like
platform() and toolchain() end up depending on the exec platform.