Prevent cyclic dependency on lcov_merger
It turns out configuration_field("coverage", "output_generator") always
has to return None if --collect_code_coverage is false, even if
--coverage_output_generator is set. If you don't, then you end up with
cyclic dependencies on lcov_merger.

We solve this by letting this configuration field point to an alias that
uses a config setting to suppress the value if coverage is not enabled.
While there, move the implementation of configuration_field() from Go to
Starlark. This should make it easier to modify (and hopefully eventually
deorbit).