Let attr.label(allow_files) use glob() NFAs
To make glob() work we added a nifty package for converting the provided
set of patterns to an NFA. In the case of attr.label(allow_files) we
essentially need to do the same thing: match pathnames. However, there
we use a completely different structure. Let's unify this by storing an
NFA there as well.