Embed a configuration reference in model_starlark_pb.File
Right now we only include some abbreviated identifier in there, but
that's not sufficient if we want to instruct the analysis to actually
build the file.

While there, also change the code to put a base64 copy of the
configuration reference in the pathname. This will make it easier for
users to figure out which configuration was used to build a file.