Let TargetCompletion actually request the contents of output files
Now that analysis seems to be mostly complete, let's start working
towards doing actual builds. The first step is to add a new function
that can be called to gain access to a single output file. This function
will need to call ConfiguredTarget to figure out which action generates
the file and invoke that.