Add a schema for recording the outputs of targets
We're going to use a B-tree, where outputs are sorted by filename. For
each of these we need to record the type, but also the source of the
file. In case of functions like ctx.actions.write() we can directly
embed a FileContents in here, allowing the file to be placed in an input
root directly.