Add a schema for encoding Args objects
Right now ctx.actions.run() throws away the command line arguments that
it receives. We should obviously save these if we want to make execution
work. Rules may call ctx.actions.args() to construct Args objects. These
can be used to delay the actual computation of arguments to execution.

In our case we'll give each action a B-tree of Args objects. In theory
it's possible to share Args objects between actions. However, this
doesn't seem to happen that frequently. Let's therefore not bother
detaching Args objects from the actions in which they are used. We'll
rely on object deduplication to clean that up.