Add an experimental tool for inspecting objects in storage
This tool is sort of the equivalent of bb_browser, except that it
currently does not spend any effort rendering objects in a user friendly
way. It simply dumps objects in JSON form. However, it does make
following links between objects easy, as it uses the field options to
determine the type of the linked object.