Store the initialization vector in the parent object
Right now we store the initialization vector we use for encrypting
objects before the payload. This is problematic, because it means that
if you know the encryption key, you can do a full scan of the CAS and
attempt to decrypt all objects. Especially for workers that need to get
access to the key that's used to encrypt files that's problematic.

This change solves this by storing the initialization vector in the
payload of the parent object. Even though we could extend 'message
Reference' to include a special field for that, I'm adding a special
DecodableReference that wraps the former, containing a "decoding
parameters" field. In addition to composing better on the Go code side,
it has the advantage that you can now refer to objects in such a way
that access to its contents is not granted explicitly.

On the Go code side we need to propagate these decoding parameters
along. In some but not all places where we use CreatedObject[T], we also
need to store it. Similarly, in some but not all places where we use
references we also need to store it. Add a Decodable[T] that can be used
to staple decoding parameters onto arbitrary types.

In places where we want to display decodable references, we now use a
JWT inspired format: "${reference}.${decodingParameters}", where both
components are base64url encoded.