Make @@rules_java++toolchains+local_jdk work
It contains a symlink farm pointing to the local Java installation.
Those symlinsk need to be expended. In other words, we need to instruct
the repo platform workers to copy such files from the installation into
the CAS.

This change adds a RepoPlatformHostPath function that's capable of doing
exactly that. When the provided path corresponds to a regular file, no
additional work needs to be performed. If it is a directory, we need to
recursively invoke RepoPlatformHostPath to get rid of any additional
directories.