Let the bonanza_bazel CLI emit clickable links
This change reworks the bonanza_bazel CLI to use a custom string
formatting library that can generate OSC 8 clickable links. This we can
use to make the build specification and stack frames link to
bonanza_browser.

In its current form, this change breaks client-side displaying of
message contents of keys in stack frames. However, I want to move the
generation for those entirely to the client side, so that we can perform
syntax highlighting and can emit additional links to bonanza_browser for
references contained within.