Handle empty build_setting_default for label settings consistently
As an extension, our implementation allows build_setting_default for
label settings to be set to None. At the Protobuf level we use an empty
string instead. We should make sure that if we restore such values, we
always map them to the None value as well.

This fixes an issue where the configuration of targets would list
superfluous build settings, as setting them to None would not remove
them from the configuration.