Extend structs to support computed fields
Types like LibraryToLink needs to provide fields that are not hashable,
such as functions and lists. However, it also needs to be hashable.
Solve this by providing an extension, allowing structs to contain
computed fields. When accessed, we call into a function that computes
the actual value using the explicitly set struct fields.