Split off TargetAction into TargetAction{Command,InputRoot}
Let's make TargetActionResult a very simple function that just calls
into two helpers to create the command and input root. The advantage of
of this approach is that we can now compute the shape of actions without
actually invoking them.

This will be useful if we want to provide tooling for running actions
locally. It also has the advantage that command and input root
generation can now happen in parallel, and can also be cached
separately.