Don't announce is_sibling_repository_layout=True
It turns out that the "sibling repository layout" is still somewhat
different from what we do. Namely, it causes the working directory of
actions to also be set to the current repo. This is not what we do.

Setting this to false causes the C++ rules to pass the correct include
path to build actions.