Make ctx.actions.declare_shareable_artifact() work somewhat better
It looks like (almost?) all places where this function is called, it is
used to declare output files underneath the current package. Callers
merely use this function, because they depend on the shareable aspect of
it (i.e., potentially having multiple targets that declare the same
outputs). This is something that Bonanza already supports for any type
of output.