Make ctx.actions.symlink(target_file=...) work properly
We were creating symlinks that still had a placeholder as a target path.
With this change applied, it now points to the file that is referenced.
Whereas Bazel emits symlinks with absolute targets, we create ones with
relative targets. This makes the resulting directories safe for use in
input roots.