Add better "reference equality" for depsets
In our case it's a bit annoying to provide reference equality, for the
reason that depsets get written to/read from storage. We may duplicate
depsets, but may also collapse identical instances.

As we should try to remain compatible with Bazel in this regard, give
every depset an identifier. For empty depsets we keep it empty, as all
empty depsets are equal to each other. For non-empty depsets we number
them sequentially, using a cryptographic hash of the current function's
key as the initial offset.

This is all suboptimal, but probably the best we can do for now. In the
future we should investigate whether there are better things we can do
here. Maybe we should just .to_list() the depset when any attempt is
made to hash/compare it?