Implement ctx.runfiles() entirely in Starlark
Right now this function calls into a native implementation that lives in
ConfiguredTarget. This one returns runfiles objects that are implemented
in Go.

This change gets rid of that, replacing runfiles objects with a simple
struct. The merge() and merge_all() methods are provided in the form of
computed fields.