Don't distinguish between default_runfiles and data_runfiles
The Bazel documentation is fairly explicit that people should avoid
using data_runfiles. ctx.runfiles() should be called with runfiles=...,
and DefaultInfo.default_runfiles should be used to access them.

Let's just make it so that our copy of DefaultInfo only has
default_runfiles, which contains the union of all runfiles that are
provided to ctx.runfiles(). We support data_runfiles in the form of a
computed field.