Extend File to be usable for directory expansion
DirectoryExpander.expand() needs to return instances of File that have a
tree relative path attached to them. An easy way to support this would
be to extend the File Protobuf message to have a tree_relative_path
field, but that doesn't seem correct to me.

Instances of File with a tree relative path never make it into storage,
as they only exist while TargetActionCommand runs. We therefore store it
as a separate field in the Starlark File object.