Add the ability to test BaseComputer
Right now none of the code under pkg/model/analysis has any unit testing
coverage. I intentionally waited with that, because I first wanted the
overall design to stabilize a bit. Now that we're pretty close to having
all of bb-storage build via Bonanza, it is time to start writing tests.

This change adds a baseComputerTester type that can be instantiated by
all tests in this package. It configured BaseComputer in such a way that
both input and output objects are stored in the form of
CreatedObjectTree. This means that BaseComputer can run without any
central storage.

As an example, we add testing coverage for some basic parts of FileRoot.