Allow inlinedtree.Build() to spill large lists into separate objects
inlinedtree.Build() can currently only store messages in separate
objects. This is problematic, because we also have many cases in which
we want to store lists of objects separately.

To achieve this, let's add a Marshalable type to model_core. We can use
this to encode all the different strategies of how we marshal data.
Right now there are only implementations for plain messages and lists of
messages. Maybe later we also want to provide an implementation for
blobs, so that we can store large Starlark strings/bytes in separate
objects.

When storing lists externally, we need to emit parent nodes. For this we
already have btree.ParentNodeComputer. Let's reuse this here. Provide a
btree.MaybeMergeNodes() function that acts as a bridge between
inlinedtree and btree. Because inlinedtree.ParentAppender does not allow
returning errors, change btree.ParentNodeComputer to also not allow it.

I'm not entirely happy with how friendly the resulting API is to use.
Let's see if we can improve it at some point in the future. For example,
I don't like how we need to pass the capturer, encoder, and parent node
computer around separately, even though I don't think you ever want
those to differ. We should really address that at some point in the
nearby future.