Let FileRoot follow symlinks into input roots
Actions are permitted to yield symlinks that point to source files or
outputs of dependencies. This means that if TargetActionResult may
report dangling symlinks. In that case we need to invoke
TargetActionInputRoot and continue resolution there.

As such dependencies are scarce, we should make sure to not invoke
TargetActionInputRoot if not needed.