Port LocalBlobAccess over to Bonanza
What is nice about Bonanza is that we only need to store objects up to
2 MB in size. This has several advantages:

- We can get rid of the data integrity cache. This mostly just existed
  to permit fast random access to large files in the case of FUSE reads.
  ParsedObjectPool now takes care of that.

- We no longer need to stream objects from disk. Because of that, we
  don't really require locking to ensure blobs are not overwritten while
  being read. We can now detect overwrites simply by doing checksum
  validation. This allows us to get rid of BlockAllocator. The
  location-blob map is now a simple ring buffer.

There are still several TODOs in the code that we need to address. First
and foremost, refreshing isn't triggered properly yet, even though the
configuration options for it exist. We also lack Prometheus metrics on
worst-case retention. These will be addressed in the future.