Add a "lossy map"
The key-location map that Buildbarn's LocalBlobAccess has is a pretty
useful data structure. It's essentially a fixed size self-cleaning map
structure. As part of Bonanza's storage backends we also want to have
something similar.

This change brings in the key-location map code from Buildbarn, but
reworks it to be generic. The API is the same, except for one minor
detail: callers can now pass through data for determining which records
in the map are expired. The LocalBlobAccess key-location map only
receives this when the record array is constructed, which isn't very
flexible.