Add lease tracking to bonanza_storage_shard
Our disk based storage backend (which is still a stub) currently needs
to deal with storing objects and tracking leases. However in our case
it's sufficient to just have a decorator that does the lease tracking
for us.

Solve this by adding a new FlatStore interface, which is essentially a
simplified version of Store that uses a new FlatReference type.
FlatReference is similar to an REv2 digest, only storing a hash and
size. We then add an adapter on top of FlatStore that does the lease
tracking.

The next step will be to add a proper FlatStore implementation that
writes objects to disk. This can be a lot simpler than LocalBlobAccess,
because we don't need to support FindMissingBlobs() and only need to
store objects up to 2 MB in size.