Let lease renewing return UploadObjectIncomplete properly
It looks like the current implementation can never return
UploadObjectIncomplete, even if some of the children of an object
actually turn out to be missing. This means that if graphs in storage
turn out to be incomplete, the system ends up in a stuck state, making
it impossible to get the graph uploaded again.