Allow remote execution clients to provide Ed25519 public keys
We're only interested in using ECDH, meaning that X25519 is sufficient.
Even though RFC 8410 defines a method for placing an X25519 public key
in an X.509 certificate, support for this seems mediocre at best. For
example, OpenSSL is not able to generate such certificates.

Simplify all of this by allowing clients to provide certificates
containing Ed25519 public keys. RFC 7748 provides a process for
converting Ed25519 keys to X25519 counterparts.