Generalize remote execution to support arbitrary elliptic curves
Right now we only support X25519/Ed25519. However, this can be
problematic if there already is certificate management infrastructure
present that uses other algorithms, such as ECDSA.

This change changes most of the existing code to no longer mention
X25519 specifically. Furthermore, on the scheduler side we now keep
track of one verification key per curve type. When a worker calls
Synchronize() and we need to return a verification response, we return
all verification public keys belonging to curves that the caller uses.