Decompose the remote worker client
Just like the remote execution client, make it so that the base
implementation is only capable of sending byte slices across. We move
the Protobuf message handling into a decorator for the Executor.

This does require us to make a minor change to the API of Executor, as
it should now be possible for the Executor to reject actions as well. We
solve this by adding an additional 'error' return value.