Rebuild bonanza_worker on top of model_executewithstorage
bonanza_builder has already been adjusted to make use of this protocol.
Now let's give bonanza_worker the same treatment. As this means we can
now use native reference types as part of model_command_pb.Action, we
can use model_filesystem_pb.DirectoryReference for the input root. This
is consistent with the rest of the code base.