Let workers translate DownloadObject() NOT_FOUND to FAILED_PRECONDITION
This is something that Buildbarn already does for REv2, but it looks
like we're going to need this as well. For our fetching worker we don't
want to return NOT_FOUND in case any object referenced by its action
does not exist. We want to limit that return code to just 404s.