Move fetching of files via HTTP out of bonanza_builder
The idea is that we can run bonanza_builder in environments that are
completely isolated from the network. It should be possible to run it in
the cloud in a closed off network. In order to fetch files belonging to
module extensions, the analyzer should call into a helper process. This
helper process can be run inside of a company network that has access to
services like GitHub/GitLab, Artifactory, etc.

This change adds a new bonanza_fetcher process that does the
downloading, which we now spin up as part of the demo deployment. We
then rewrite the HttpFileContents analysis function to issue a remote
execution client request, just like how we run build actions.