Ensure calling the remote execution client uploads actions
Right now we have this implicit guarantee that actions are automatically
uploaded to the cache prior to uploading. This is because actions are
computed in one analysis function and invoked in the other. However, if
action creation and invocation happen in the same function, the action
may be missing.

Let's fix this by calling UploadDAG() as part of our decorator for the
execution client.