Change the golden test for file Merkle tree creation to use ChaCha8
The Mersenne Twister implementation we're currently using has a Read()
implementation that discards data when the amount of data that is read
is not a multiple of eight. This means that our golden test isn't very
reproducible.

As the Go standard library has in the meantime gained an acceptable
ChaCha8 implementation, let's just use that.