Make registration of predeclared platform queues work
This code had a panic("TODO") in it, because I hadn't spent any time yet
to implement the registration of new platform public keys outside of
Synchronize(). Shuffle a bunch of types and functions around, so that we
can call this from within RegisterPredeclaredPlatformQueue().

While there, change RegisterPredeclaredPlatformQueue() to accept a list
of blobs instead of ecdh.PublicKey. The scheduler shouldn't enforce a
canonical representation of public keys. The original public key should
be retained at all costs.