Require that target patterns provided by the client are apparent
Just like with stock Bazel, the client should take the working directory
into account to determine which targets need to be built. This means
that the client already needs to interpret and normalize these patterns.
It's therefore no additional effort for the client to guarantee that
target patterns are apparent.